Validate required Java version in `misc` subproject.
Previously, validation of required version was done during loading of
build definition. This was needlessly restricting because not everything
in scalariform requires particular Java version. After closer inspection,
it turns out that only the `misc` subproject depends on Swing and cares
about the specific Java version.

We refactor validation logic to run right before `compile` task in `misc`
subproject. That makes it possible to ignore this subproject and build the
rest of scalariform on any Java version.