Make scalariform compatible with 2.12
Lots of changes were required. I couldn't compile the tests because of
lots of changes in scalatest, therefore I just skipped running the
tests. Making the sbt build compatible with 2.12 wasn't required
actually because there is a maven build. It can be run like this:

    mvn -Pscala-2.12.x clean install -DskipTests

Afterwards, we need to upload the stuff in

    scalariform.update/target/site/

to S3.